Fixed compiler warnings.

+++ common/lc_math.h
@@ -3,10 +3,10 @@
 #include <math.h>
 #include <float.h>
 
-#define LC_DTOR (M_PI / 180.0)
-#define LC_RTOD (180.0 / M_PI)
-#define LC_PI   (M_PI)
-#define LC_2PI  (2.0 * M_PI)
+#define LC_DTOR (static_cast<float>(M_PI / 180))
+#define LC_RTOD (static_cast<float>(180 / M_PI))
+#define LC_PI (static_cast<float>(M_PI))
+#define LC_2PI (static_cast<float>(2 * M_PI))
 
 #define LC_RGB(r,g,b) LC_RGBA(r,g,b,255)
 #define LC_RGBA(r,g,b,a) ((lcuint32)(((lcuint8) (r) | ((lcuint16) (g) << 8)) | (((lcuint32) (lcuint8) (b)) << 16) | (((lcuint32) (lcuint8) (a)) << 24)))
